app/requirements.txt
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
colorama==0.4.6
coloredlogs==15.0.1
cryptography==45.0.6
customtkinter==5.2.2
darkdetect==0.8.0
diffusers==0.35.1
filelock==3.19.1
flatbuffers==25.2.10
fsspec==2025.7.0
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.2
onnxruntime==1.22.1
onnxruntime-directml==1.22.0
packaging==25.0
pillow==11.3.0
protobuf==6.32.0
pycparser==2.22
pyreadline3==3.5.4
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.5
safetensors==0.6.2
setuptools==80.9.0
sympy==1.14.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
zipp==3.23.0
